add config base and credential
