+++ lib/deploygate.rb
@@ -30,7 +30,8 @@ require "deploygate/commands/logout"
 require "deploygate/commands/deploy"
 require "deploygate/commands/deploy/push"
 require "deploygate/commands/deploy/build"
-require "deploygate/config"
+require "deploygate/config/base"
+require "deploygate/config/credential"
 require "deploygate/session"
 require "deploygate/deploy"
 require "deploygate/build"
