+++ spec/deploygate/config/base_spec.rb
@@ -1,4 +1,8 @@
-describe DeployGate::Config do
+describe DeployGate::Config::Base do
+  before do
+    allow(DeployGate::Config::Base).to receive(:file_path).and_return(File.join(SPEC_FILE_PATH, 'test_files/.dg/base'))
+  end
+
   describe "#write" do
     it "write data" do
       write_data = {
@@ -6,9 +10,9 @@ describe DeployGate::Config do
           :token => 'token'
       }
       allow(File).to receive(:open).and_return(StringIO.new("", "w+"))
-      DeployGate::Config.write(write_data)
+      DeployGate::Config::Base.write(write_data)
 
-      file = File.open(DeployGate::Config.file_path)
+      file = File.open(DeployGate::Config::Base.file_path)
       expect(file.string).to eq(write_data.to_json.to_s)
     end
   end
@@ -20,7 +24,7 @@ describe DeployGate::Config do
           :token => 'token'
       }.to_json.to_s
       allow(File).to receive(:open).and_return(StringIO.new(write_data))
-      data = DeployGate::Config.read
+      data = DeployGate::Config::Base.read
 
       expect(data).to eq(JSON.parse(write_data))
     end
